requirements.txt
aiocache==0.12.2
aiohttp==3.9.1   
aiosignal==1.3.1 
attrs==23.1.0    
contourpy==1.2.0 
cycler==0.12.1   
discord.py==2.3.2
fonttools==4.46.0
frozenlist==1.4.0
idna==3.6        
kiwisolver==1.4.5
lxml==4.9.3      
matplotlib==3.8.2
multidict==6.0.4 
numpy==1.26.2    
packaging==23.2  
pandas==2.1.3
Pillow==10.1.0
pycairo==1.25.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
seaborn==0.13.0
six==1.16.0
tzdata==2023.3
yarl==1.9.3